Rename normalization_variations to norm_variations
