+++ variations/norm_variations.py
